Use abstraction for GUI classes
