+++ builder/src/main/java/rhit/presentation/SwingGui.java
@@ -1,0 +1,16 @@
+package rhit.presentation;
+
+import lombok.Setter;
+
+abstract class SwingGui {
+  @Setter
+  private static SwingGui visibleFrame;
+
+  static void showFrame() {
+    visibleFrame.show();
+  }
+
+  abstract void show();
+
+  abstract void handleContinue();
+}
